added ability to register custom log formatter (#131)
* added ability to register custom log formatter

* changed function type declaration to work on older versions of Go

* changed to an args struct per PR, refactored logging tests to make it easier to see the scenarios

* fixed formatting and linting issues

* split logging into its own file

* removed writer from the LogFormatterParams struct

* split all logging tests into a separate file

* fixed formatting
