+++ handlers_test.go
@@ -5,14 +5,11 @@
 package handlers
 
 import (
-	"bytes"
-	"net"
 	"net/http"
 	"net/http/httptest"
 	"net/url"
 	"strings"
 	"testing"
-	"time"
 )
 
 const (
@@ -70,233 +67,6 @@ func TestMethodHandler(t *testing.T) {
 		if body := rec.Body.String(); body != test.body {
 			t.Fatalf("%d: wrong body, got %q want %q", i, body, test.body)
 		}
-	}
-}
-
-func TestMakeLogger(t *testing.T) {
-	rec := httptest.NewRecorder()
-	logger := makeLogger(rec)
-	// initial status
-	if logger.Status() != http.StatusOK {
-		t.Fatalf("wrong status, got %d want %d", logger.Status(), http.StatusOK)
-	}
-	// WriteHeader
-	logger.WriteHeader(http.StatusInternalServerError)
-	if logger.Status() != http.StatusInternalServerError {
-		t.Fatalf("wrong status, got %d want %d", logger.Status(), http.StatusInternalServerError)
-	}
-	// Write
-	logger.Write([]byte(ok))
-	if logger.Size() != len(ok) {
-		t.Fatalf("wrong size, got %d want %d", logger.Size(), len(ok))
-	}
-	// Header
-	logger.Header().Set("key", "value")
-	if val := logger.Header().Get("key"); val != "value" {
-		t.Fatalf("wrong header, got %s want %s", val, "value")
-	}
-}
-
-func TestWriteLog(t *testing.T) {
-	loc, err := time.LoadLocation("Europe/Warsaw")
-	if err != nil {
-		panic(err)
-	}
-	ts := time.Date(1983, 05, 26, 3, 30, 45, 0, loc)
-
-	// A typical request with an OK response
-	req := newRequest("GET", "http://example.com")
-	req.RemoteAddr = "192.168.100.5"
-
-	buf := new(bytes.Buffer)
-	writeLog(buf, req, *req.URL, ts, http.StatusOK, 100)
-	log := buf.String()
-
-	expected := "192.168.100.5 - - [26/May/1983:03:30:45 +0200] \"GET / HTTP/1.1\" 200 100\n"
-	if log != expected {
-		t.Fatalf("wrong log, got %q want %q", log, expected)
-	}
-
-	// CONNECT request over http/2.0
-	req = &http.Request{
-		Method:     "CONNECT",
-		Proto:      "HTTP/2.0",
-		ProtoMajor: 2,
-		ProtoMinor: 0,
-		URL:        &url.URL{Host: "www.example.com:443"},
-		Host:       "www.example.com:443",
-		RemoteAddr: "192.168.100.5",
-	}
-
-	buf = new(bytes.Buffer)
-	writeLog(buf, req, *req.URL, ts, http.StatusOK, 100)
-	log = buf.String()
-
-	expected = "192.168.100.5 - - [26/May/1983:03:30:45 +0200] \"CONNECT www.example.com:443 HTTP/2.0\" 200 100\n"
-	if log != expected {
-		t.Fatalf("wrong log, got %q want %q", log, expected)
-	}
-
-	// Request with an unauthorized user
-	req = newRequest("GET", "http://example.com")
-	req.RemoteAddr = "192.168.100.5"
-	req.URL.User = url.User("kamil")
-
-	buf.Reset()
-	writeLog(buf, req, *req.URL, ts, http.StatusUnauthorized, 500)
-	log = buf.String()
-
-	expected = "192.168.100.5 - kamil [26/May/1983:03:30:45 +0200] \"GET / HTTP/1.1\" 401 500\n"
-	if log != expected {
-		t.Fatalf("wrong log, got %q want %q", log, expected)
-	}
-
-	// Request with url encoded parameters
-	req = newRequest("GET", "http://example.com/test?abc=hello%20world&a=b%3F")
-	req.RemoteAddr = "192.168.100.5"
-
-	buf.Reset()
-	writeLog(buf, req, *req.URL, ts, http.StatusOK, 100)
-	log = buf.String()
-
-	expected = "192.168.100.5 - - [26/May/1983:03:30:45 +0200] \"GET /test?abc=hello%20world&a=b%3F HTTP/1.1\" 200 100\n"
-	if log != expected {
-		t.Fatalf("wrong log, got %q want %q", log, expected)
-	}
-}
-
-func TestWriteCombinedLog(t *testing.T) {
-	loc, err := time.LoadLocation("Europe/Warsaw")
-	if err != nil {
-		panic(err)
-	}
-	ts := time.Date(1983, 05, 26, 3, 30, 45, 0, loc)
-
-	// A typical request with an OK response
-	req := newRequest("GET", "http://example.com")
-	req.RemoteAddr = "192.168.100.5"
-	req.Header.Set("Referer", "http://example.com")
-	req.Header.Set(
-		"User-Agent",
-		"Mozilla/5.0 (Macintosh; Intel Mac OS X 10_8_2) AppleWebKit/537.33 "+
-			"(KHTML, like Gecko) Chrome/27.0.1430.0 Safari/537.33",
-	)
-
-	buf := new(bytes.Buffer)
-	writeCombinedLog(buf, req, *req.URL, ts, http.StatusOK, 100)
-	log := buf.String()
-
-	expected := "192.168.100.5 - - [26/May/1983:03:30:45 +0200] \"GET / HTTP/1.1\" 200 100 \"http://example.com\" " +
-		"\"Mozilla/5.0 (Macintosh; Intel Mac OS X 10_8_2) " +
-		"AppleWebKit/537.33 (KHTML, like Gecko) Chrome/27.0.1430.0 Safari/537.33\"\n"
-	if log != expected {
-		t.Fatalf("wrong log, got %q want %q", log, expected)
-	}
-
-	// CONNECT request over http/2.0
-	req1 := &http.Request{
-		Method:     "CONNECT",
-		Host:       "www.example.com:443",
-		Proto:      "HTTP/2.0",
-		ProtoMajor: 2,
-		ProtoMinor: 0,
-		RemoteAddr: "192.168.100.5",
-		Header:     http.Header{},
-		URL:        &url.URL{Host: "www.example.com:443"},
-	}
-	req1.Header.Set("Referer", "http://example.com")
-	req1.Header.Set(
-		"User-Agent",
-		"Mozilla/5.0 (Macintosh; Intel Mac OS X 10_8_2) AppleWebKit/537.33 "+
-			"(KHTML, like Gecko) Chrome/27.0.1430.0 Safari/537.33",
-	)
-
-	buf = new(bytes.Buffer)
-	writeCombinedLog(buf, req1, *req1.URL, ts, http.StatusOK, 100)
-	log = buf.String()
-
-	expected = "192.168.100.5 - - [26/May/1983:03:30:45 +0200] \"CONNECT www.example.com:443 HTTP/2.0\" 200 100 \"http://example.com\" " +
-		"\"Mozilla/5.0 (Macintosh; Intel Mac OS X 10_8_2) " +
-		"AppleWebKit/537.33 (KHTML, like Gecko) Chrome/27.0.1430.0 Safari/537.33\"\n"
-	if log != expected {
-		t.Fatalf("wrong log, got %q want %q", log, expected)
-	}
-
-	// Request with an unauthorized user
-	req.URL.User = url.User("kamil")
-
-	buf.Reset()
-	writeCombinedLog(buf, req, *req.URL, ts, http.StatusUnauthorized, 500)
-	log = buf.String()
-
-	expected = "192.168.100.5 - kamil [26/May/1983:03:30:45 +0200] \"GET / HTTP/1.1\" 401 500 \"http://example.com\" " +
-		"\"Mozilla/5.0 (Macintosh; Intel Mac OS X 10_8_2) " +
-		"AppleWebKit/537.33 (KHTML, like Gecko) Chrome/27.0.1430.0 Safari/537.33\"\n"
-	if log != expected {
-		t.Fatalf("wrong log, got %q want %q", log, expected)
-	}
-
-	// Test with remote ipv6 address
-	req.RemoteAddr = "::1"
-
-	buf.Reset()
-	writeCombinedLog(buf, req, *req.URL, ts, http.StatusOK, 100)
-	log = buf.String()
-
-	expected = "::1 - kamil [26/May/1983:03:30:45 +0200] \"GET / HTTP/1.1\" 200 100 \"http://example.com\" " +
-		"\"Mozilla/5.0 (Macintosh; Intel Mac OS X 10_8_2) " +
-		"AppleWebKit/537.33 (KHTML, like Gecko) Chrome/27.0.1430.0 Safari/537.33\"\n"
-	if log != expected {
-		t.Fatalf("wrong log, got %q want %q", log, expected)
-	}
-
-	// Test remote ipv6 addr, with port
-	req.RemoteAddr = net.JoinHostPort("::1", "65000")
-
-	buf.Reset()
-	writeCombinedLog(buf, req, *req.URL, ts, http.StatusOK, 100)
-	log = buf.String()
-
-	expected = "::1 - kamil [26/May/1983:03:30:45 +0200] \"GET / HTTP/1.1\" 200 100 \"http://example.com\" " +
-		"\"Mozilla/5.0 (Macintosh; Intel Mac OS X 10_8_2) " +
-		"AppleWebKit/537.33 (KHTML, like Gecko) Chrome/27.0.1430.0 Safari/537.33\"\n"
-	if log != expected {
-		t.Fatalf("wrong log, got %q want %q", log, expected)
-	}
-}
-
-func TestLogPathRewrites(t *testing.T) {
-	var buf bytes.Buffer
-
-	handler := http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
-		req.URL.Path = "/" // simulate http.StripPrefix and friends
-		w.WriteHeader(200)
-	})
-	logger := LoggingHandler(&buf, handler)
-
-	logger.ServeHTTP(httptest.NewRecorder(), newRequest("GET", "/subdir/asdf"))
-
-	if !strings.Contains(buf.String(), "GET /subdir/asdf HTTP") {
-		t.Fatalf("Got log %#v, wanted substring %#v", buf.String(), "GET /subdir/asdf HTTP")
-	}
-}
-
-func BenchmarkWriteLog(b *testing.B) {
-	loc, err := time.LoadLocation("Europe/Warsaw")
-	if err != nil {
-		b.Fatalf(err.Error())
-	}
-	ts := time.Date(1983, 05, 26, 3, 30, 45, 0, loc)
-
-	req := newRequest("GET", "http://example.com")
-	req.RemoteAddr = "192.168.100.5"
-
-	b.ResetTimer()
-
-	buf := &bytes.Buffer{}
-	for i := 0; i < b.N; i++ {
-		buf.Reset()
-		writeLog(buf, req, *req.URL, ts, http.StatusUnauthorized, 500)
 	}
 }
